+++ src/gen_java/enum_.rs
@@ -3,8 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use super::{CodeType, Config, potentially_add_external_package};
-use crate::ComponentInterface;
-use uniffi_bindgen::backend::Literal;
+use uniffi_bindgen::interface::ComponentInterface;
 
 #[derive(Debug)]
 pub struct EnumCodeType {
@@ -29,18 +28,6 @@ impl CodeType for EnumCodeType {
 
     fn canonical_name(&self) -> String {
         format!("Type{}", self.id)
-    }
-
-    fn literal(&self, literal: &Literal, ci: &ComponentInterface, config: &Config) -> String {
-        if let Literal::Enum(v, _) = literal {
-            format!(
-                "{}.{}",
-                self.type_label(ci, config),
-                super::JavaCodeOracle.enum_variant_name(v)
-            )
-        } else {
-            unreachable!();
-        }
     }
 
     fn ffi_converter_instance(&self, config: &Config, ci: &ComponentInterface) -> String {