+++ src/templates/Int16Helper.java
@@ -1,32 +1,30 @@
 package {{ config.package_name() }};
 
-import java.nio.ByteBuffer;
-
-public enum FfiConverterShort implements FfiConverter<Short, Short>{
+public enum FfiConverterShort implements FfiConverter<java.lang.Short, java.lang.Short>{
   INSTANCE;
 
     @Override
-    public Short lift(Short value) {
+    public java.lang.Short lift(java.lang.Short value) {
         return value;
     }
 
     @Override
-    public Short read(ByteBuffer buf) {
+    public java.lang.Short read(java.nio.ByteBuffer buf) {
         return buf.getShort();
     }
 
     @Override
-    public Short lower(Short value) {
+    public java.lang.Short lower(java.lang.Short value) {
         return value;
     }
 
     @Override
-    public long allocationSize(Short value) {
+    public long allocationSize(java.lang.Short value) {
         return 2L;
     }
 
     @Override
-    public void write(Short value, ByteBuffer buf) {
+    public void write(java.lang.Short value, java.nio.ByteBuffer buf) {
         buf.putShort(value);
     }
 }